fix: separated redis config
